+++ KlayGE/Plugins/Src/Render/OpenGLES/OGLESRenderEngine.cpp
@@ -687,7 +687,7 @@ namespace KlayGE
 		caps_.max_simultaneous_rts = 1;
 
 		glGetIntegerv(GL_MAX_VERTEX_ATTRIBS, &temp);
-		caps_.max_vertex_streams = temp;
+		caps_.max_vertex_streams = static_cast<uint8_t>(temp);
 
 		caps_.hw_instancing_support = true;
 		caps_.instance_id_support = false;
